1.完成了用户账户登录与验证

+++ app/src/main/java/com/xianyu/yixian_client/Model/Tcp/CustomHeartbeatHandler.java
@@ -36,7 +36,7 @@ public class CustomHeartbeatHandler extends ChannelHandlerAdapter {
             ClientResponseModel respond = (ClientResponseModel) msg;
             ClientRequestModel request = socketClient.tasks.get(Integer.parseInt((respond.Id)));
             if(request != null){
-                socketClient.tasks.remove(Integer.parseInt((respond.Id)));
+                socketClient.tasks.remove(Long.parseLong((respond.Id)));
                 request.setResult(respond.Result);
             }
             else {
